refactor: rename _template_validation.py to _rubric_kind_validation.py to clarify scope
