+++ src/karenina/schemas/entities/_rubric_kind_validation.py
